Moved files to knowledge base package
